geometry: document VertexStreamView lifetime, fix type width, squash warning
- Document that VertexStreamView::data pointers are valid only while
  the source Mesh is alive and unmodified (immediate-consumption API)
- Fix uint32_t vs size_t loop variable in BooleanOperation.cpp
- Fix remaining -Wunused-result warning in test_MeshDiagnosticOverlay

+++ apps/Nexus-Modeling/src/kernel/src/geometry/BooleanOperation.cpp
@@ -277,7 +277,7 @@ void computeBooleanResult(const std::vector<Vec3>& posA, const std::vector<Trian
             // Simple pointer comparison for determinism
             uint32_t oldIdx = 0;
             float minDist = std::numeric_limits<float>::max();
-            for (uint32_t i = 0; i < posA.size(); ++i) {
+            for (size_t i = 0; i < posA.size(); ++i) {
                 float d = vec3Length(vec3Sub(posA[i], p));
                 if (d < minDist) {
                     minDist = d;
